+++ CHANGELOG.md
@@ -7,6 +7,9 @@ V0.2.2 - Not released yet
 * Add browser action: ◢ + C = Copy trailing segment of URL
 * Change: Use aosd-cat instead of notify-send for notifications
 * Improve Gnome Screenshot action ◣ + V: Set feh window always on top
+* Remove Evolution Mail actions
+* Add Claws Maiil action: ◢ + H = Toggle "Hide read threads"
+* Change action: ◢ + M = Launch/bring to front Claws Mail
 
 V0.2.1 - 2020/12/13
 -------------------